[Comgr][hotswap] Make comgr-object-utils self-contained for the transpile driver
comgr-metadata.cpp and comgr-symbol.cpp reference parseTargetIdentifier,
setCStr, and the DataSymbol ctor/dtor, whose definitions sat in comgr.cpp.
amd_comgr links comgr.cpp so it never noticed, but the transpile driver links
only the comgr-object-utils objects, so on MSVC those references became
unresolved externals. Relying on the linker to strip the unused functions that
carry the references (the way --gc-sections does elsewhere) did not hold up
under MSVC's /OPT:REF.

Move the definitions into the object-utils translation units that already use
them -- parseTargetIdentifier alongside the metadata helpers it calls, setCStr
and DataSymbol into comgr-symbol.cpp -- so the object set the driver links is
closed under its own references. comgr.cpp keeps calling them and resolves
against object-utils exactly as before, with no duplicate definitions.
