+++ amd/comgr/src/comgr.cpp
@@ -238,69 +238,8 @@ bool COMGR::isDataKindValid(amd_comgr_data_kind_t DataKind) {
          DataKind <= AMD_COMGR_DATA_KIND_LAST;
 }
 
-amd_comgr_status_t COMGR::setCStr(char *&Dest, StringRef Src, size_t *Size) {
-  free(Dest);
-  Dest = reinterpret_cast<char *>(malloc(Src.size() + 1));
-  if (!Dest) {
-    return AMD_COMGR_STATUS_ERROR_OUT_OF_RESOURCES;
-  }
-  memcpy(Dest, Src.data(), Src.size());
-  Dest[Src.size()] = '\0';
-  if (Size) {
-    *Size = Src.size();
-  }
-  return AMD_COMGR_STATUS_SUCCESS;
-}
-
 StringRef COMGR::getComgrHashIdentifier() {
   return xstringify(AMD_COMGR_VERSION_ID);
-}
-
-amd_comgr_status_t COMGR::parseTargetIdentifier(StringRef IdentStr,
-                                                TargetIdentifier &Ident) {
-  // TODO: Only use AMDGPU::TargetID for this parsing.
-  SmallVector<StringRef, 5> IsaNameComponents;
-  IdentStr.split(IsaNameComponents, '-', 4);
-  if (IsaNameComponents.size() != 5) {
-    return AMD_COMGR_STATUS_ERROR_INVALID_ARGUMENT;
-  }
-
-  Ident.Arch = IsaNameComponents[0];
-  Ident.Vendor = IsaNameComponents[1];
-  Ident.OS = IsaNameComponents[2];
-  Ident.Environ = IsaNameComponents[3];
-
-  Ident.Features.clear();
-  IsaNameComponents[4].split(Ident.Features, ':');
-
-  // The first ':'-separated token is the processor; it may be empty when the
-  // ISA is encoded in an "amdgpu<subarch>" arch instead.
-  Ident.Processor = Ident.Features[0];
-  Ident.Features.erase(Ident.Features.begin());
-
-  if (IdentStr == "spirv64-amd-amdhsa--amdgcnspirv" ||
-      IdentStr == "spirv64-amd-amdhsa-unknown-amdgcnspirv") {
-    // Features not supported for SPIR-V
-    if (!Ident.Features.empty())
-      return AMD_COMGR_STATUS_ERROR_INVALID_ARGUMENT;
-    return AMD_COMGR_STATUS_SUCCESS;
-  }
-
-  // Validate the target ID and resolve the canonical processor.
-  size_t IsaIndex;
-  amd_comgr_status_t Status =
-      metadata::getIsaIndex(IdentStr, IsaIndex, &Ident.Processor);
-  if (Status != AMD_COMGR_STATUS_SUCCESS) {
-    return Status;
-  }
-
-  for (auto Feature : Ident.Features) {
-    if (!metadata::isSupportedFeature(IsaIndex, Feature)) {
-      return AMD_COMGR_STATUS_ERROR_INVALID_ARGUMENT;
-    }
-  }
-
-  return AMD_COMGR_STATUS_SUCCESS;
 }
 
 void COMGR::ensureLLVMInitialized() {
@@ -490,9 +429,6 @@ std::string DataMeta::convertDocNodeToString(msgpack::DocNode DocNode) {
   }
   return DocNode.toString();
 }
-
-DataSymbol::DataSymbol(SymbolContext *DataSym) : DataSym(DataSym) {}
-DataSymbol::~DataSymbol() { delete DataSym; }
 
 amd_comgr_status_t AMD_COMGR_API
     // NOLINTNEXTLINE(readability-identifier-naming)
